test remove armv7 i386
